frontend: Use relative paths in index.html
Use relative paths in index.html for loading global.css, bundle.css,
bundle.js and favicon.png so that website can be easily rendered in a
subdirectory structure as is the default for GitHub pages or GCP storage
buckets. We will use auto-deployment to storage buckets in a follow up
step.

While at it move bundle.[js|css] from public/build to public/ to keep
things nice and flat. This comes at the cost of combining generated and
manually created content, we will see how we'll go.